% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\alias{detect.character}
\alias{detect.data.frame}
\alias{detect.factor}
\alias{detect.numeric}
\title{Detect if a data object contains PII.}
\usage{
detect(.x)

\method{detect}{character}(.x)

\method{detect}{factor}(.x)

\method{detect}{numeric}(.x)

\method{detect}{data.frame}(.x)
}
\arguments{
\item{.x}{A data object.}
}
\value{
A logical value indicating if that data object contains PII.
}
\description{
Detect if a data object contains PII.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Method for character vectors.

\item \code{factor}: Method for factor vectors.

\item \code{numeric}: Method for numeric vectors.

\item \code{data.frame}: Method for data.frames.
}}
\examples{
# Examples
}

