\name{detector-package}
\alias{detector-package}
\alias{detector}
\docType{package}
\title{\packageTitle{detector}}
\description{\packageDescription{detector}}
\section{Package Content}{\packageIndices{detector}}
\author{\packageAuthor{detector}}
\section{Maintainer}{\packageMaintainer{detector}}
\keyword{package}
