% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detector.R
\docType{package}
\name{detector}
\alias{detector}
\alias{detector-package}
\title{detector: Detect Data Containing Personally Identifiable Information}
\description{
detector: Detect Data Containing Personally Identifiable Information
}

