% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/email_addresses.R
\name{has_email_addresses}
\alias{has_email_addresses}
\title{Test if a character vector has any e-mail addresses.}
\usage{
has_email_addresses(.x)
}
\arguments{
\item{.x}{A character vector.}
}
\value{
A logical value indicating if that string has any e-mail addresses.
}
\description{
Test if a character vector has any e-mail addresses.
}
\examples{
# Examples
}

