% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/national_identification_numbers.R
\name{has_national_identification_numbers}
\alias{has_national_identification_numbers}
\title{Test if a character vector has any national identification numbers.}
\usage{
has_national_identification_numbers(.x)
}
\arguments{
\item{.x}{A character vector.}
}
\value{
A logical value indicating if that string has any national identification numbers.
}
\description{
Test if a character vector has any national identification numbers.
}
\examples{
# Examples
}

