% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/email_addresses.R
\name{is_email_address}
\alias{is_email_address}
\title{Test if a string is an e-mail address.}
\usage{
is_email_address(.x)
}
\arguments{
\item{.x}{A character vector.}
}
\value{
A logical value indicating if that string is an e-mail address.
}
\description{
Test if a string is an e-mail address.
}
\examples{
# Examples
is_email_address("hello") # FALSE
is_email_address("hello@world.edu") # TRUE
}

