% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/national_identification_numbers.R
\name{is_national_identification_number}
\alias{is_national_identification_number}
\title{Test if a string is a national identification number.}
\usage{
is_national_identification_number(.x)
}
\arguments{
\item{.x}{A string or numeric vector.}
}
\value{
A logical value indicating if that string is a national identification number.
}
\description{
Strictly works for only US national identification numbers.
}
\examples{
# Examples
is_national_identification_number("hello") # FALSE
is_national_identification_number(65884) # FALSE
is_national_identification_number("111-33-5555") # TRUE
is_national_identification_number(1113335555) # FALSE
}

