% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phone_numbers.R
\name{is_phone_number}
\alias{is_phone_number}
\title{Test if a string is a phone number.}
\usage{
is_phone_number(.x)
}
\arguments{
\item{.x}{A string or numeric vector.}
}
\value{
A logical value indicating if that string is a phone number.
}
\description{
Strictly works for only US phone numbers.
}
\examples{
# Examples
is_phone_number("hello") # FALSE
is_phone_number(65884) # FALSE
is_phone_number("111-333-5555") # TRUE
is_phone_number(1113335555) # TRUE
}

