% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/social_security_numbers.R
\name{is_social_security_number}
\alias{is_social_security_number}
\title{Test if a string is a social security number.}
\usage{
is_social_security_number(.x)
}
\arguments{
\item{.x}{A string or numeric vector.}
}
\value{
A logical value indicating if that string is a social security number.
}
\description{
Strictly works for only US social security numbers.
}
\examples{
# Examples
is_social_security_number("hello") # FALSE
is_social_security_number(65884) # FALSE
is_social_security_number("111-333-5555") # TRUE
is_social_security_number(1113335555) # TRUE
}

