% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/phone_numbers.R
\name{has_phone_numbers}
\alias{has_phone_numbers}
\title{Test if a character vector has any phone numbers.}
\usage{
has_phone_numbers(.x)
}
\arguments{
\item{.x}{A character vector.}
}
\value{
A logical value indicating if that string has any phone numbers.
}
\description{
Test if a character vector has any phone numbers.
}
\examples{
# Examples
}

